chore(test): adopt co-located unit-test convention + drop tests/unit
Pick a single convention: tests live next to their subject as
<file>.test.ts, not under tests/unit/. Co-location surfaces missing
tests at a glance, keeps tests with their subject when files move,
and shortens the relative imports. CLAUDE.md + docs/TESTING.adoc
are explicit that this is the only supported pattern now.

Migration: split tests/unit/nip17SubjectInterop.test.ts (the only
test the repo had) into three co-located files matching the subjects
each describe block exercised:
  - src/services/nostrService.test.ts        (outgoing kind-14)
  - src/utils/nip17Unwrap.test.ts            (incoming subject parser)
  - src/utils/syntheticGroupId.test.ts       (room-key derivation)

The third file imports across both src/services and src/utils to
exercise the wire-format roundtrip — co-located next to the API
whose contract the assertion checks.

jest.config.js testMatch tightened to <rootDir>/src/**/*.test.{ts,tsx}
only. tests/unit/ directory removed.
