+++ src/services/nostrService.test.ts
@@ -1,0 +1,49 @@
+/**
+ * Wire-format guards for `createGroupChatRumor` (outgoing NIP-17
+ * kind-14). The subject tag is what foreign clients (Amethyst /
+ * Quartz, 0xchat) read to display the group name — Lightning
+ * Piggy's outgoing messages MUST include it for cross-client
+ * interop. Issue #271.
+ */
+
+import { createGroupChatRumor } from './nostrService';
+
+const PK_A = 'a'.repeat(64);
+const PK_B = 'b'.repeat(64);
+const PK_C = 'c'.repeat(64);
+
+describe('createGroupChatRumor (outgoing kind-14)', () => {
+  it('includes a subject tag carrying the group name', () => {
+    const rumor = createGroupChatRumor({
+      senderPubkey: PK_A,
+      subject: 'Pizza Friday',
+      memberPubkeys: [PK_B, PK_C],
+      content: 'who is in?',
+    });
+    const subject = rumor.tags.find((t) => t[0] === 'subject');
+    expect(subject).toEqual(['subject', 'Pizza Friday']);
+  });
+
+  it('emits one p tag per recipient member', () => {
+    const rumor = createGroupChatRumor({
+      senderPubkey: PK_A,
+      subject: 'x',
+      memberPubkeys: [PK_B, PK_C],
+      content: 'hi',
+    });
+    const ps = rumor.tags.filter((t) => t[0] === 'p').map((t) => t[1]);
+    expect(ps).toEqual([PK_B, PK_C]);
+  });
+
+  it('builds a kind-14 rumor (NIP-17 chat)', () => {
+    const rumor = createGroupChatRumor({
+      senderPubkey: PK_A,
+      subject: 'x',
+      memberPubkeys: [PK_B],
+      content: 'hi',
+    });
+    expect(rumor.kind).toBe(14);
+    expect(rumor.pubkey).toBe(PK_A);
+    expect(rumor.content).toBe('hi');
+  });
+});
