+++ src/utils/nip17Unwrap.test.ts
@@ -1,0 +1,39 @@
+/**
+ * Wire-format guards for `subjectFromRumor` (incoming NIP-17 kind-14
+ * subject-tag parser). Used by `tryRouteGroupRumor` to materialise
+ * synthetic groups from foreign-client messages (Amethyst, 0xchat).
+ * Issue #271.
+ */
+
+import { subjectFromRumor, type DecodedRumor } from './nip17Unwrap';
+
+const PK_A = 'a'.repeat(64);
+const PK_B = 'b'.repeat(64);
+
+describe('subjectFromRumor (incoming kind-14)', () => {
+  function rumorWithTags(tags: string[][]): DecodedRumor {
+    return {
+      pubkey: PK_A,
+      created_at: 1,
+      kind: 14,
+      tags,
+      content: '',
+    };
+  }
+
+  it('returns the subject value when present', () => {
+    expect(subjectFromRumor(rumorWithTags([['subject', 'Hello']]))).toBe('Hello');
+  });
+
+  it('returns null when there is no subject tag', () => {
+    expect(subjectFromRumor(rumorWithTags([['p', PK_B]]))).toBeNull();
+  });
+
+  it('returns null when subject is whitespace-only', () => {
+    expect(subjectFromRumor(rumorWithTags([['subject', '   ']]))).toBeNull();
+  });
+
+  it('trims surrounding whitespace', () => {
+    expect(subjectFromRumor(rumorWithTags([['subject', '  Pizza Friday  ']]))).toBe('Pizza Friday');
+  });
+});
